pyproject.toml
[project]
name = "diagram-agent"
version = "0.1.0"
description = "Conversational AI agent for generating draw.io architecture diagrams"
requires-python = ">=3.12"
dependencies = [
    "strands-agents>=1.0.0",
    "strands-agents-tools>=0.1.0",
    "lxml>=5.0.0",
    "boto3>=1.34.0",
    "graphviz>=0.20",
    "jsonpatch>=1.33",
]

[project.optional-dependencies]
dev = [
    "pytest>=8.0.0",
    "bedrock-agentcore-starter-toolkit",
]
